Fix(Pages): Settings & Download pages fixed
I changed the settings page from admin-settings to settings and i also updated the view and the urls to reflect the changes Consequently the view has been changed from admin-settings to settings same as the urls on the django. I also updated the download page to the new design. I fix the logout showing at the bottom of the sidebar.